Moved code under the serviceme namespace
This makes is a lot more specific, rather than sitting around at the top
level com.github.saem top level. Further clean-up is required, but
started adding initial unit tests.
