--- src/main/java/com/github/saem/greeting/Greeter.java
@@ -1,5 +1,0 @@
-package com.github.saem.greeting;
-
-public final class Greeter {
-    public final String phrase = "Hello, World!";
-}
