--- src/main/java/com/github/saem/scaffolding/ApplicationConfig.java
@@ -1,3 +1,0 @@
-package com.github.saem.scaffolding;
-
-public abstract class ApplicationConfig {}
