--- src/main/java/com/github/saem/scaffolding/components/ComponentConfig.java
@@ -1,3 +1,0 @@
-package com.github.saem.scaffolding.components;
-
-public interface ComponentConfig {}
