--- src/main/java/com/github/saem/scaffolding/components/InitializedComponent.java
@@ -1,5 +1,0 @@
-package com.github.saem.scaffolding.components;
-
-public interface InitializedComponent {
-    InitializedComponent start();
-}
