--- src/main/java/com/github/saem/scaffolding/components/healthchecks/HealthCheckConfig.java
@@ -1,6 +1,0 @@
-package com.github.saem.scaffolding.components.healthchecks;
-
-import com.github.saem.scaffolding.components.ComponentConfig;
-
-public final class HealthCheckConfig implements ComponentConfig {
-}
