--- src/main/java/com/github/saem/scaffolding/components/metrics/MetricsComponent.java
@@ -1,18 +1,0 @@
-package com.github.saem.scaffolding.components.metrics;
-
-import com.codahale.metrics.MetricRegistry;
-import com.github.saem.scaffolding.components.Component;
-import com.github.saem.scaffolding.components.InitializedComponent;
-
-public class MetricsComponent implements Component<MetricsConfig>,
-        InitializedComponent {
-    public final MetricRegistry registry = new MetricRegistry();
-
-    @Override
-    public MetricsComponent initialize(
-            final MetricsConfig componentConfig
-    ) { return this; }
-
-    @Override
-    public MetricsComponent start() { return this; }
-}
