--- src/main/java/com/github/saem/scaffolding/components/metrics/MetricsConfig.java
@@ -1,6 +1,0 @@
-package com.github.saem.scaffolding.components.metrics;
-
-import com.github.saem.scaffolding.components.ComponentConfig;
-
-public class MetricsConfig implements ComponentConfig {
-}
