+++ src/main/java/com/github/saem/serviceme/greeting/Greeter.java
@@ -1,0 +1,5 @@
+package com.github.saem.serviceme.greeting;
+
+public final class Greeter {
+    public final String phrase = "Hello, World!";
+}
