+++ src/main/java/com/github/saem/serviceme/scaffolding/ApplicationConfig.java
@@ -1,0 +1,3 @@
+package com.github.saem.serviceme.scaffolding;
+
+public abstract class ApplicationConfig {}
