+++ src/main/java/com/github/saem/serviceme/scaffolding/InvalidHostException.java
@@ -1,0 +1,7 @@
+package com.github.saem.serviceme.scaffolding;
+
+public final class InvalidHostException extends RuntimeException {
+    public InvalidHostException(final String invalidHost) {
+        super("This is an invalid host: " + invalidHost);
+    }
+}
