+++ src/main/java/com/github/saem/serviceme/scaffolding/components/ComponentConfig.java
@@ -1,0 +1,3 @@
+package com.github.saem.serviceme.scaffolding.components;
+
+public interface ComponentConfig {}
