+++ src/main/java/com/github/saem/serviceme/scaffolding/components/InitializedComponent.java
@@ -1,0 +1,5 @@
+package com.github.saem.serviceme.scaffolding.components;
+
+public interface InitializedComponent {
+    InitializedComponent start();
+}
