+++ src/main/java/com/github/saem/serviceme/scaffolding/components/healthchecks/HealthCheckConfig.java
@@ -1,0 +1,6 @@
+package com.github.saem.serviceme.scaffolding.components.healthchecks;
+
+import com.github.saem.serviceme.scaffolding.components.ComponentConfig;
+
+public final class HealthCheckConfig implements ComponentConfig {
+}
