+++ src/main/java/com/github/saem/serviceme/scaffolding/components/metrics/MetricsConfig.java
@@ -1,0 +1,6 @@
+package com.github.saem.serviceme.scaffolding.components.metrics;
+
+import com.github.saem.serviceme.scaffolding.components.ComponentConfig;
+
+public class MetricsConfig implements ComponentConfig {
+}
