+++ src/test/java/com/github/saem/serviceme/server/GreeterComponentsTest.java
@@ -1,0 +1,5 @@
+package com.github.saem.serviceme.server;
+
+public final class GreeterComponentsTest {
+
+}
